==> notebooks/0_data_preprocess.ipynb <==
#### Notes

This script includes

- Selecting countries
- Labelling data
- Checking data types

**Don't run if raw data not existed**

##### See **README.md** for data sources

library(tidyverse)

countries <- c("India", "South Africa", "Kenya", "Brazil")

# Risk profile

raw <- read_csv("data_raw/Table-6.3.2.csv")
head(raw)

risk <- raw %>% 
    filter(Country %in% countries) %>%
    mutate(
        Und_all_a = as.numeric(Und_all_a),
        HIV_1549_a = as.numeric(HIV_1549_a),
        HIV_1549_a = ifelse(Country == "India", 0.2, HIV_1549_a),
        Smoking_15_f = as.numeric(Smoking_15_f),
        Smoking_15_m = as.numeric(Smoking_15_m),
        across(Und_all_a:Alc_15_m, function(x) x / 100)
    )
risk

risk %>% write_csv("data/gho_risk.csv")

### TB incidence disaggregated

raw <- read_csv("data_raw/TB_burden_age_sex_2023-03-02.csv")

head(raw)

inc_by_risk <- raw %>% 
    filter(country %in% countries) %>%
    rename(Country = country, M = best, L = lo, U = hi)

head(inc_by_risk)

inc_by_risk %>% write_csv("data/who_inc.csv")

### Population size

raw <- read_csv("data_raw/WPP2022.csv")

pop <- raw %>%
    filter(Location %in% countries) %>%
    filter(Time %in% 2020:2030) %>%
    filter(Variant == "Medium") %>%
    select(
        iso3 = ISO3_code, iso2 = ISO2_code, Country = Location, Year = Time,
        Age = AgeGrp, PopMale, PopFemale, PopTotal      
    ) %>%
    mutate(
        Age = ifelse(Age == "100+", "100", Age),
        Age = as.numeric(Age),
        PopMale = PopMale * 1000,
        PopFemale = PopFemale * 1000,
        PopTotal = PopTotal * 1000
    )

head(pop)

pop %>% write_csv("data/wpp_pop.csv")
