==> notebooks/1_synthesise_pop.ipynb <==
from risk import Group, fill_tags
import pandas as pd
import json
import os

def build_population(sel_pop):
    pop = list()

    for _, row in sel_pop.iterrows():
        age = row.Age
        pop.append(Group(age, 'm', row['PopMale']))
        pop.append(Group(age, 'f', row['PopFemale']))

    return pop    

gho_risk = pd.read_csv('data/gho_risk.csv')
wpp_pop = pd.read_csv('data/wpp_pop.csv')

gho_risk.head()

gho_risk = {k: dict(v) for k, v in gho_risk.set_index('Country').iterrows()}

wpp_pop.head()

### Example run

country = 'India'

sel_risk = gho_risk[country]

sel_risk = [
    ('Und_all_a', 'Und', sel_risk['Und_all_a'], lambda gp: True), 
    ('HIV_1549_a', 'HIV', sel_risk['HIV_1549_a'], lambda gp: (50 > gp.Age >= 15)), 
    ('Smoking_15_f', 'Smk', sel_risk['Smoking_15_f'], lambda gp: (gp.Age >= 15) & (gp.Sex == 'f')), 
    ('Smoking_15_m', 'Smk', sel_risk['Smoking_15_m'], lambda gp: (gp.Age >= 15) & (gp.Sex == 'm')), 
    ('Dia_18_f', 'Dia', sel_risk['Dia_18_f'], lambda gp: (gp.Age >= 18) & (gp.Sex == 'f')), 
    ('Dia_18_m', 'Dia', sel_risk['Dia_18_m'], lambda gp: (gp.Age >= 18) & (gp.Sex == 'm')), 
    ('Alc_15_f', 'Alc', sel_risk['Alc_15_f'], lambda gp: (gp.Age >= 15) & (gp.Sex == 'f')), 
    ('Alc_15_m', 'Alc', sel_risk['Alc_15_m'], lambda gp: (gp.Age >= 15) & (gp.Sex == 'm'))
]


sel_risk

sel_pop = wpp_pop[wpp_pop.Country == country]
sel_pop = sel_pop[sel_pop.Year == 2021]
sel_pop

pop0 = build_population(sel_pop)

def print_pop(pop):
    print('N:', sum([gp.N for gp in pop]))
    
    for key, tag, p_targets, fil in sel_risk:
        eligible = [gp for gp in pop if fil(gp)]
        n_eligible = sum([gp.N for gp in eligible])
        n_risk = sum([gp.N for gp in eligible if tag in gp.Tags])
    
        print('--', key,  tag, f'Targets: {p_targets: .1%}, Filled: {n_risk / n_eligible:.1%}')
    

pop_independent = fill_tags(pop0, sel_risk, fn_rank = lambda gp: 1)

print_pop(pop_independent)

pop_exclusive = fill_tags(pop0, sel_risk, fn_rank = lambda gp: - len(gp.Tags))

print_pop(pop_exclusive)

pop_cluster = fill_tags(pop0, sel_risk, fn_rank = lambda gp: len(gp.Tags))

print_pop(pop_cluster)


n00, n01, n10, n11 = 0, 0, 0, 0
tag1, tag2 = 'Dia', 'Smk'
for gp in pop_cluster:
    if tag1 in gp.Tags:
        if tag2 in gp.Tags:
            n11 += gp.N
        else:
            n10 += gp.N
    else:
        if tag2 in gp.Tags:
            n01 += gp.N
        else:
            n00 += gp.N

print(True, True, n11)
print(True, False, n10)
print(False, True, n01)
print(False, False, n00)

#### Run

countries = list(gho_risk.keys())
countries

for country in countries:
    os.makedirs(f'data/{country}', exist_ok=True)
    
    sel_risk = gho_risk[country]
    sel_risk = [
        ('Und_all_a', 'Und', sel_risk['Und_all_a'], lambda gp: True), 
        ('HIV_1549_a', 'HIV', sel_risk['HIV_1549_a'], lambda gp: (50 > gp.Age >= 15)), 
        ('Smoking_15_f', 'Smk', sel_risk['Smoking_15_f'], lambda gp: (gp.Age >= 15) & (gp.Sex == 'f')), 
        ('Smoking_15_m', 'Smk', sel_risk['Smoking_15_m'], lambda gp: (gp.Age >= 15) & (gp.Sex == 'm')), 
        ('Dia_18_f', 'Dia', sel_risk['Dia_18_f'], lambda gp: (gp.Age >= 18) & (gp.Sex == 'f')), 
        ('Dia_18_m', 'Dia', sel_risk['Dia_18_m'], lambda gp: (gp.Age >= 18) & (gp.Sex == 'm')), 
        ('Alc_15_f', 'Alc', sel_risk['Alc_15_f'], lambda gp: (gp.Age >= 15) & (gp.Sex == 'f')), 
        ('Alc_15_m', 'Alc', sel_risk['Alc_15_m'], lambda gp: (gp.Age >= 15) & (gp.Sex == 'm'))
    ]
    
    
    sel_pop = wpp_pop[wpp_pop.Country == country]
    sel_pop = sel_pop[sel_pop.Year == 2021]
    pop0 = build_population(sel_pop)

    js = [gp.to_dict() for gp in pop0]
    pd.DataFrame(js).fillna(0).to_csv(f'data/{country}/pop.csv')
    json.dump(js, open(f'data/{country}/pop.json', 'w'))
    
    js = [gp.to_dict() for gp in pop_independent]
    pd.DataFrame(js).fillna(0).to_csv(f'data/{country}/pop_independent.csv')
    json.dump(js, open(f'data/{country}/pop_independent.json', 'w'))
    
    pop_independent = fill_tags(pop0, sel_risk, fn_rank = lambda gp: 1)
    js = [gp.to_dict() for gp in pop_independent]
    pd.DataFrame(js).fillna(0).to_csv(f'data/{country}/pop_independent.csv')
    json.dump(js, open(f'data/{country}/pop_independent.json', 'w'))
    
    pop_exclusive = fill_tags(pop0, sel_risk, fn_rank = lambda gp: - len(gp.Tags))
    js = [gp.to_dict() for gp in pop_exclusive]
    pd.DataFrame(js).fillna(0).to_csv(f'data/{country}/pop_exclusive.csv')
    json.dump(js, open(f'data/{country}/pop_exclusive.json', 'w'))
    
    pop_cluster = fill_tags(pop0, sel_risk, fn_rank = lambda gp: len(gp.Tags))
    js = [gp.to_dict() for gp in pop_cluster]
    pd.DataFrame(js).fillna(0).to_csv(f'data/pop_cluster.csv')
    json.dump(js, open(f'data/{country}/pop_cluster.json', 'w'))

os.makedirs(f'data/{country}', exist_ok=True)
